Update storybook to use chromatic, as well as cleaning up code a bit more
