+++ package.json
@@ -1,8 +1,13 @@
 {
   "name": "ntcore-react",
   "description": "React bindings for ntcore-ts-client.",
-  "author": "Josh Bacon",
-  "repository": "https://github.com/CrispyBacon1999/ntcore-react",
+  "author": {
+    "name": "Josh Bacon"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/CrispyBacon1999/ntcore-react.git"
+  },
   "keywords": [
     "frc",
     "networktables",
@@ -18,7 +23,8 @@
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
     "storybook": "start-storybook -p 6006",
-    "build:storybook": "build-storybook -o docs-build"
+    "build:storybook": "build-storybook -o docs-build",
+    "chromatic": "chromatic --exit-zero-on-changes"
   },
   "dependencies": {
     "ntcore-ts-client": "^0.2.1",
@@ -37,6 +43,7 @@
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@vitejs/plugin-react-swc": "^3.0.0",
+    "chromatic": "^6.17.4",
     "jest": "^29.4.2",
     "jest-dom": "^4.0.0",
     "jest-environment-jsdom": "^29.4.2",
@@ -70,5 +77,11 @@
   "files": [
     "dist",
     "readme.md"
-  ]
+  ],
+  "bugs": {
+    "url": "https://github.com/CrispyBacon1999/ntcore-react/issues"
+  },
+  "readme": "ERROR: No README data found!",
+  "homepage": "https://github.com/CrispyBacon1999/ntcore-react#readme",
+  "_id": "ntcore-react@1.1.0"
 }
