+++ src/index.ts
@@ -3,5 +3,6 @@ export { default as useNTState } from "./lib/useNTState";
 export { default as useNTValue } from "./lib/useNTValue";
 export { default as useNTConnected } from "./lib/useNTConnected";
 export { type default as NTTopicType } from "./lib/NTTopicType";
+export { default as Units } from "./lib/util/Units";
 
 export * from './components';
